# Conflicts:
#	create.php
#	read.php

+++ delete.php
@@ -1,9 +1,6 @@
 <?php
 
-include('controller/MainController.php');
-echo (new MainController())->delete();
-
-/*if (isset($_GET['id'])) {
+if (isset($_GET['id'])) {
 
     $Connect2DB = new PDO("pgsql:host=localhost; dbname=postgres", "postgres", "gfnhbjn", array(\PDO::MYSQL_ATTR_INIT_COMMAND => 'SET NAMES utf8'));
     $DeleteFromDB = $Connect2DB -> prepare("DELETE FROM article WHERE id=:id;");
